Tweaked episode info sheet background interaction
This is a sort-of workaround for an apparent bug, where using .enabled
without upThrough seems to do nothing if there is only a single detent.
In this case I'm currently using two detents so this isn't a problem,
but I'm making the fix now in case I ever decide to change that.