Restored HTML parsing of TVmaze synopses
I read somewhere that NSAttributedString must run on the main thread
when parsing HTML or it crashes. This might be the cause of the crashes
I found before. It's running on the main thread now and it's working.

Formatting is problematic. By default it comes up in a small Times New
Roman. To impose a base font, a kludge from StackOverflow is adopted.
The main alternative is to include the font style in the HTML, but using
the system font is a little iffy from HTML so I went this way.

Because HTML parsing must run on the main thread, the operation is now
async and the view is modified accordingly.